update autograder node to initialize mode and tool position, needs to change mode after the node is up

+++ cg_ws/src/planning/include/planning/autograder_node.hpp
@@ -35,11 +35,9 @@ private:
   /* Parameters */  
   float slip_thresh_;
   float half_deadband_;
-  int design_blade_pos_;
-  int max_des_blade_pos_; 
+  double design_blade_pos_;
+  double max_des_blade_pos_; 
   
-  
-
 }; // class AutoGraderNode
 
 } // namespace planning
